requirements.txt
# Core dependencies
fastapi
uvicorn
websockets>=12.0
pydantic
python-dotenv
openai
twilio
requests

# Database
sqlalchemy
asyncpg
psycopg2-binary